+++ scripts/txt2img.py
@@ -240,7 +240,7 @@ def main():
     model = load_model_from_config(config, f"{opt.ckpt}")
 
     device = torch.device("cuda") if torch.cuda.is_available() else torch.device("cpu")
-    model = model.to(device)
+    model = model.to(device).half()
 
     if opt.plms:
         sampler = PLMSSampler(model)
